requirements.txt
openai
langchain
python-dotenv
langchain-openai